# docs/openapi.yml
openapi: 3.0.3
info:
  title: Awexen LMS API
  version: 1.0.0
  description: Authoritative API for the Awexen learning platform.
paths:
...
  /api/v1/learning/enrollments/{enrollment_id}/:
    get:
      operationId: learning_enrollment_detail
      parameters:
      - in: path
        name: enrollment_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - api
      security:
      - bearerAuth: []
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT